refactor(issue): move open issue to separate file
